+++ src/RedisClient/Command/Traits/Version4x0/CommandsTrait.php
@@ -14,7 +14,6 @@ use RedisClient\Command\Traits\AbstractCommandsTrait;
 use RedisClient\Command\Traits\Version2x8\LatencyCommandsTrait;
 use RedisClient\Command\Traits\Version2x8\PubSubCommandsTrait;
 use RedisClient\Command\Traits\Version3x0\ClusterCommandsTrait;
-use RedisClient\Command\Traits\Version2x8\ConnectionCommandsTrait;
 use RedisClient\Command\Traits\Version2x8\HyperLogLogCommandsTrait;
 use RedisClient\Command\Traits\Version2x6\ListsCommandsTrait;
 use RedisClient\Command\Traits\Version3x0\SortedSetsCommandsTrait;
@@ -22,7 +21,6 @@ use RedisClient\Command\Traits\Version2x6\TransactionsCommandsTrait;
 use RedisClient\Command\Traits\Version3x2\GeoCommandsTrait;
 use RedisClient\Command\Traits\Version3x2\HashesCommandsTrait;
 use RedisClient\Command\Traits\Version3x2\ScriptingCommandsTrait;
-use RedisClient\Command\Traits\Version3x2\ServerCommandsTrait;
 use RedisClient\Command\Traits\Version3x2\SetsCommandsTrait;
 use RedisClient\Command\Traits\Version3x2\StringsCommandsTrait;
 
@@ -38,6 +36,7 @@ trait CommandsTrait {
     use KeysCommandsTrait;
     use LatencyCommandsTrait;
     use ListsCommandsTrait;
+    use MemoryCommandsTrait;
     use PubSubCommandsTrait;
     use ScriptingCommandsTrait;
     use ServerCommandsTrait;