+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 ## CHANGELOG
 
 ### v1.6.0 (2016-12-24)
+- Added support for Redis 4.0 (the Client was tested with Redis 4.0 RC2).
 - Added support for Redis Cluster.
-- Client was tested with Redis 4.0 RC1
+- Added command **SWAPDB**, **UNLINK**, **MEMORY** for Redis >= 4.0
+- Updated command **FLUSHALL**, **FLUSHDB** for Redis >= 4.0
 
 ### v1.5.1 (2016-08-17)
 - Fixed critical bug: https://github.com/cheprasov/php-redis-client/pull/45 Thanks to @BrianFranklin for help.